{"schema":"decodex/commit/1","summary":"Modularize Rust host FFI ABI payload owners","authority":"manual"}
